+++ CHANGELOG.md
@@ -139,6 +139,8 @@ version 0.2.0, thus covering only a fraction of the project’s history.
   output file to target for a given tweak etc.
     - Central housing, a new feature adding a body separate from the main body.
     - Custom bodies.
+    - Flanges: Arbitrarily positioned screws for connecting custom bodies to
+      their parent bodies.
 - The `by-key` section of parameters has become more flexible.
     - It is now possible to influence all keys on a row, without duplicating
       settings for all columns in that row.